fix(collapse): move generateUUID utility to tedi/helpers and update imports #866
