+++ libs/angular-components/community/components/form/radio/radio/radio.component.ts
@@ -15,7 +15,7 @@ import {
   ComponentInputs,
   FeedbackTextComponent,
 } from "@tehik-ee/tedi-angular/tedi";
-import { generateUUID } from "../../../../../utils/generateUUID";
+import { generateUUID } from "../../../../../tedi/helpers/generateUUID";
 
 export type RadioValue = string;
