+++ libs/angular-components/tedi/components/buttons/collapse/collapse.component.ts
@@ -11,7 +11,7 @@ import {
 import { TediTranslationPipe } from "../../../services/translation/translation.pipe";
 import { IconComponent } from "../../base/icon/icon.component";
 import { TextComponent } from "../../base/text/text.component";
-import { generateUUID } from "../../../../utils/generateUUID";
+import { generateUUID } from "../../../helpers/generateUUID";
 
 export type ArrowType = "default" | "secondary";
