+++ libs/angular-components/tedi/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from "./generateUUID";
