+++ libs/angular-components/tedi/index.ts
@@ -2,3 +2,4 @@ export * from "./components";
 export * from "./directives";
 export * from "./services";
 export * from "./types";
+export * from "./helpers";
