+++ libs/angular-components/utils/index.ts
@@ -1,2 +1,1 @@
 export * from "./createBreakpointArgTypes";
-export * from "./generateUUID";
